+++ parser/src/processors/filekey_processor.rs
@@ -15,29 +15,29 @@ pub fn collect_filekey(
     last_char: String,
 ) {
     let clone_parser = parser.clone();
-    if let parser::Collecting::FileKey(ref mut filekeydata) = parser.current {
-        if !filekeydata.keyname_collected {
+    if let parser::Collecting::FileKey(ref mut file_key_data) = parser.current {
+        if !file_key_data.key_name_collected {
             let current_reliability = utils::reliable_name_range(
                 utils::ReliableNameRanges::VariableName,
                 letter_char.to_string(),
             );
 
             if current_reliability.reliable {
-                if filekeydata.data.keyname.is_empty() {
-                    filekeydata.data.keyname_location.range_start = parser.pos;
+                if file_key_data.data.key_name.is_empty() {
+                    file_key_data.data.key_name_location.range_start = parser.pos;
                 }
 
-                if (last_char == " " || last_char == "\n") && filekeydata.data.keyname != "" {
-                    filekeydata.data.keyname = letter_char.to_string();
+                if (last_char == " " || last_char == "\n") && file_key_data.data.key_name != "" {
+                    file_key_data.data.key_name = letter_char.to_string();
                 }
 
-                filekeydata.data.keyname += letter_char;
-                filekeydata.data.keyname_location.range_end = parser.pos;
+                file_key_data.data.key_name += letter_char;
+                file_key_data.data.key_name_location.range_end = parser.pos;
             } else {
-                if letter_char == "=" && !filekeydata.data.keyname.is_empty() {
-                    filekeydata.keyname_collected = true;
+                if letter_char == "=" && !file_key_data.data.key_name.is_empty() {
+                    file_key_data.key_name_collected = true;
                 } else if letter_char != " "
-                    && (letter_char == "@" && filekeydata.data.keyname != "")
+                    && (letter_char == "@" && file_key_data.data.key_name != "")
                 {
                     errors.push(error::Error {
                         scope: "filekey_processor".to_string(),
@@ -59,9 +59,9 @@ pub fn collect_filekey(
                     });
                 }
             }
-        } else if letter_char == ";" && filekeydata.data.value.is_type_complete() {
+        } else if letter_char == ";" && file_key_data.data.value.is_type_complete() {
             if clone_parser
-                .check_keyword(filekeydata.data.keyname.clone())
+                .check_keyword(file_key_data.data.key_name.clone())
                 .found
             {
                 errors.push(error::Error {
@@ -74,21 +74,21 @@ pub fn collect_filekey(
                         error::errorList::error_s24.message.clone(),
                         vec![error::ErrorBuildField {
                             key: "token".to_string(),
-                            value: filekeydata.data.keyname.clone(),
+                            value: file_key_data.data.key_name.clone(),
                         }],
                     ),
                     pos: defs::Cursor {
-                        range_start: filekeydata.data.keyname_location.range_start,
-                        range_end: filekeydata
+                        range_start: file_key_data.data.key_name_location.range_start,
+                        range_end: file_key_data
                             .data
-                            .keyname_location
+                            .key_name_location
                             .range_end
                             .clone()
                             .skip_char(1),
                     },
                 });
             }
-            if utils::is_reserved(&filekeydata.data.keyname) {
+            if utils::is_reserved(&file_key_data.data.key_name) {
                 errors.push(error::Error {
                     scope: "filekey_processor".to_string(),
                     debug_message: "replace_filekey_processor83".to_string(),
@@ -99,27 +99,27 @@ pub fn collect_filekey(
                         error::errorList::error_s21.message.clone(),
                         vec![error::ErrorBuildField {
                             key: "token".to_string(),
-                            value: filekeydata.data.keyname.clone(),
+                            value: file_key_data.data.key_name.clone(),
                         }],
                     ),
                     pos: defs::Cursor {
-                        range_start: filekeydata.data.keyname_location.range_start,
-                        range_end: filekeydata
+                        range_start: file_key_data.data.key_name_location.range_start,
+                        range_end: file_key_data
                             .data
-                            .keyname_location
+                            .key_name_location
                             .range_end
                             .clone()
                             .skip_char(1),
                     },
                 });
             }
-            filekeydata.value_collected = true;
+            file_key_data.value_collected = true;
             parser.collected.push(parser.current.clone());
             parser.current = parser::Collecting::None;
         } else {
             let mut will_be_itered = syntax::variable::VariableCollector {
                 data: syntax::variable::Variable {
-                    value: filekeydata.data.value.clone(),
+                    value: file_key_data.data.value.clone(),
                     ..Default::default()
                 },
                 ..Default::default()
@@ -137,12 +137,12 @@ pub fn collect_filekey(
                 errors.extend(itered_filekey_vector.errors);
             }
 
-            if filekeydata.data.value_location.is_zero() {
-                filekeydata.data.value_location.range_start = parser.pos;
+            if file_key_data.data.value_location.is_zero() {
+                file_key_data.data.value_location.range_start = parser.pos;
             }
 
-            filekeydata.data.value = itered_filekey_vector.itered_data.data.value;
-            filekeydata.data.value_location.range_end = parser.pos;
+            file_key_data.data.value = itered_filekey_vector.itered_data.data.value;
+            file_key_data.data.value_location.range_end = parser.pos;
         }
     }
 }