+++ parser/src/syntax/condition.rs
@@ -5,6 +5,7 @@ use serde::Serialize;
 
 use crate::alloc::string::String;
 use crate::alloc::vec::Vec;
+use crate::alloc::boxed::Box;
 
 #[derive(PartialEq, Debug, Clone, Serialize)]
 pub enum ConditionType {
@@ -22,7 +23,7 @@ impl Default for ConditionType {
 #[derive(PartialEq, Debug, Clone, Default, Serialize)]
 pub struct ConditionChain {
     pub rtype: ConditionType,
-    pub condition: types::cloak_type::CloakType,
+    pub condition: Box<types::Types>,
     pub inside_code: Vec<Collecting>,
 }