+++ parser/src/syntax/definers.rs
@@ -55,6 +55,11 @@ pub struct CollectiveType {
     pub has_key: bool,
 }
 
+#[derive(PartialEq, Debug, Clone, Serialize, Default)]
+pub struct NullableType {
+    pub value: Box<DefinerCollecting>,
+}
+
 #[derive(PartialEq, Debug, Clone, Serialize, EnumAsInner)]
 pub enum DefinerCollecting {
     Array(ArrayType),
@@ -63,6 +68,7 @@ pub enum DefinerCollecting {
     Function(FunctionType),
     Cloak(CloakType),
     Collective(CollectiveType),
+    Nullable(NullableType),
     Dynamic,
 }
 
@@ -86,6 +92,13 @@ impl DefinerCollecting {
                 DefinerCollecting::GrowableArray(data) => {
                     if let DefinerCollecting::GrowableArray(other_data) = other {
                         other_data.rtype.same_as(*data.rtype)
+                    } else {
+                        false
+                    }
+                }
+                DefinerCollecting::Nullable(data) => {
+                    if let DefinerCollecting::Nullable(other_data) = other {
+                        other_data.value.same_as(*data.value)
                     } else {
                         false
                     }
@@ -151,6 +164,7 @@ impl DefinerCollecting {
         match self {
             DefinerCollecting::Array(data) => !data.complete,
             DefinerCollecting::GrowableArray(data) => !data.complete,
+            DefinerCollecting::Nullable(data) => data.value.is_type_empty(),
             DefinerCollecting::Generic(data) => data.rtype.is_empty(),
             DefinerCollecting::Function(data) => !data.complete,
             DefinerCollecting::Cloak(data) => !data.complete,
@@ -163,6 +177,7 @@ impl DefinerCollecting {
         match self {
             DefinerCollecting::Array(data) => data.complete,
             DefinerCollecting::GrowableArray(data) => data.complete,
+            DefinerCollecting::Nullable(data) => data.value.is_definer_complete(),
             DefinerCollecting::Generic(data) => !data.rtype.is_empty(),
             DefinerCollecting::Function(data) => data.complete,
             DefinerCollecting::Cloak(data) => data.complete,
@@ -175,6 +190,7 @@ impl DefinerCollecting {
         match self {
             DefinerCollecting::Array(_) => false,
             DefinerCollecting::GrowableArray(_) => false,
+            DefinerCollecting::Nullable(_) => false,
             DefinerCollecting::Generic(_) => true,
             DefinerCollecting::Function(_) => false,
             DefinerCollecting::Cloak(_) => false,
@@ -187,6 +203,7 @@ impl DefinerCollecting {
         match self {
             DefinerCollecting::Array(_) => "array".to_string(),
             DefinerCollecting::GrowableArray(_) => "dynamicArray".to_string(),
+            DefinerCollecting::Nullable(_) => "nullAble".to_string(),
             DefinerCollecting::Generic(data) => data.rtype.clone(),
             DefinerCollecting::Function(_) => "function".to_string(),
             DefinerCollecting::Cloak(_) => "cloak".to_string(),