+++ parser/src/syntax/file_key.rs
@@ -5,15 +5,15 @@ use serde::Serialize;
 
 #[derive(PartialEq, Debug, Clone, Default, Serialize)]
 pub struct FileKey {
-    pub keyname: String,
+    pub key_name: String,
     pub value: types::Types,
-    pub keyname_location: defs::Cursor,
+    pub key_name_location: defs::Cursor,
     pub value_location: defs::Cursor,
 }
 
 #[derive(PartialEq, Debug, Clone, Default, Serialize)]
 pub struct FileKeyCollector {
     pub data: FileKey,
-    pub keyname_collected: bool,
+    pub key_name_collected: bool,
     pub value_collected: bool,
 }