+++ parser/src/syntax/types/collective_type.rs
@@ -8,7 +8,7 @@ use alloc::vec::Vec;
 
 #[derive(PartialEq, Default, Debug, Clone, Serialize)]
 pub struct CollectiveEntry {
-    pub key: String,
+    pub key: Box<types::Types>,
     pub value: Box<types::Types>,
     pub key_pos: defs::Cursor,
     pub value_pos: defs::Cursor,
@@ -16,7 +16,7 @@ pub struct CollectiveEntry {
 
 #[derive(PartialEq, Default, Debug, Clone, Serialize)]
 pub struct CollectiveEntryCollector {
-    pub data: Box<types::Types>,
+    pub data: CollectiveEntry,
     pub key_collected: bool,
     pub value_collected: bool,
 }
@@ -30,4 +30,20 @@ pub struct Collective {
 pub struct CollectiveCollector {
     pub complete: bool,
     pub data: Collective,
+}
+
+impl CollectiveCollector {
+    pub fn has_dedup(&self) -> bool {
+        let mut existent_names: Vec<types::Types> = Vec::with_capacity(self.data.entries.len());
+        let mut duplicate = false;
+        for i in &self.data.entries {
+            if existent_names.contains(&*i.data.key.clone()) {
+                duplicate = true;
+                break;
+            } else {
+                existent_names.push(*i.data.key.clone())
+            }
+        }
+        duplicate
+    }
 }