+++ tests/syntax_tests.rs
@@ -2,16 +2,17 @@ mod syntax {
     mod arithmetic_operators;
     mod array;
     mod bool;
-    mod variable;
     mod char;
+    mod collective;
     mod comparison_operators;
     mod float;
     mod function;
+    mod import;
     mod integer;
     mod logical_operators;
     mod negative;
-    mod string;
-    mod import;
     mod ret;
+    mod string;
+    mod variable;
     mod variable_value;
 }