+++ CHANGELOG.md
@@ -32,6 +32,6 @@
 # v2.7.0
   - Parser
     * Comment bug fix
-    * Collective introduced 
+    * Collective introduced
     * Negative introduced (!!true) 
-    * Support more
+    * Nullable types introduced